+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the Audibene Code Snippets will be documented here.
 - Updated Classes in Advanced Mode 
 - Included `if`, `if empty`, `if not empty`, `if equals`, `if not equals`, `if else`, `else` in Aura
 - Variant and Size attributes added in aura and lwc icon snippets
+- SLDS Table Classes added
 
 ## Amber - v1.0.2 
 - 10 New Snippets added for JavaScript